Refactor test structure.
