+++ .github/workflows/06-junit-functional-python.yaml
@@ -10,7 +10,7 @@ on: [push]
 jobs:
   build:
     runs-on: ubuntu-latest
-    timeout-minutes: 30
+    timeout-minutes: 25
 
     steps:
       - uses: actions/checkout@v2
