+++ src/test/java/cloud/filibuster/functional/java/suppress_combinations/JUnitFilibusterHelloPartialHelloExternalGrpcWithSuppressCombinationsTest.java
@@ -1,4 +1,4 @@
-package cloud.filibuster.functional.java_hello.nondeterministic;
+package cloud.filibuster.functional.java.suppress_combinations;
 
 import cloud.filibuster.examples.Hello;
 import cloud.filibuster.examples.HelloServiceGrpc;
@@ -31,7 +31,7 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
  * Test simple annotation usage.
  */
 @TestMethodOrder(MethodOrderer.OrderAnnotation.class)
-public class JUnitFilibusterDataNondeterminismTest extends JUnitBaseTest {
+public class JUnitFilibusterHelloPartialHelloExternalGrpcWithSuppressCombinationsTest extends JUnitBaseTest {
     private final static Set<String> testExceptionsThrown = new HashSet<>();
 
     private static int numberOfTestsExecuted = 0;
@@ -39,7 +39,7 @@ public class JUnitFilibusterDataNondeterminismTest extends JUnitBaseTest {
     private static int numberOfExceptionsThrown = 0;
 
     @DisplayName("Test partial hello server grpc route with Filibuster. (MyHelloService, MyWorldService)")
-    @FilibusterTest(serverBackend=FilibusterLocalServerBackend.class, dataNondeterminism=true, maxIterations=10)
+    @FilibusterTest(serverBackend=FilibusterLocalServerBackend.class, suppressCombinations=true, maxIterations=30)
     @Order(1)
     public void testMyHelloAndMyWorldServiceWithFilibuster() throws InterruptedException {
         ManagedChannel helloChannel = ManagedChannelBuilder
@@ -52,53 +52,94 @@ public class JUnitFilibusterDataNondeterminismTest extends JUnitBaseTest {
         numberOfTestsExecuted++;
 
         HelloServiceGrpc.HelloServiceBlockingStub blockingStub = HelloServiceGrpc.newBlockingStub(helloChannel);
-        Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian " + Math.random()).build();
+        Hello.HelloRequest request = Hello.HelloRequest.newBuilder().setName("Armerian").build();
 
         try {
-            Hello.HelloReply reply = blockingStub.partialHello(request);
-            assertTrue(reply.getMessage().contains("Hello, Armerian"));
+            Hello.HelloReply reply = blockingStub.partialHelloExternalGrpc(request);
+            assertEquals("Hello, Hello, Hello, Armerian!!", reply.getMessage());
             assertFalse(wasFaultInjected());
         } catch (Throwable t) {
             numberOfExceptionsThrown++;
+            testExceptionsThrown.add(t.getMessage());
 
             boolean wasFaultInjected = wasFaultInjected();
 
-            if (wasFaultInjected) {
-                testExceptionsThrown.add(t.getMessage());
+            boolean firstRPCFailed = false;
+            boolean secondRPCFailed = false;
 
+            if (wasFaultInjected) {
+                // First RPC failed.
                 if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
                     expected = true;
+                    firstRPCFailed = true;
                 }
 
                 if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
                     expected = true;
+                    firstRPCFailed = true;
                 }
 
                 if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
                     expected = true;
+                    firstRPCFailed = true;
                 }
 
                 if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
                     expected = true;
+                    firstRPCFailed = true;
                 }
 
-                boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("WorldService");
-                assertTrue(wasFaultInjectedOnWorldService);
+                if (firstRPCFailed) {
+                    boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("WorldService");
+                    assertTrue(wasFaultInjectedOnWorldService);
 
-                boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod("cloud.filibuster.examples.WorldService/World");
-                assertTrue(wasFaultInjectedOnWorldMethod);
+                    boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod("cloud.filibuster.examples.WorldService/World");
+                    assertTrue(wasFaultInjectedOnWorldMethod);
 
-                boolean wasFaultInjectedOnRequest = wasFaultInjectedOnRequest(request.toString());
-                assertTrue(wasFaultInjectedOnRequest);
+                    boolean wasFaultInjectedOnRequest = wasFaultInjectedOnRequest(request.toString());
+                    assertTrue(wasFaultInjectedOnRequest);
 
-                boolean wasFaultInjectedOnWorldMethodWithPayload = wasFaultInjectedOnMethodWherePayloadContains("cloud.filibuster.examples.WorldService/World", request.toString());
-                assertTrue(wasFaultInjectedOnWorldMethodWithPayload);
+                    boolean wasFaultInjectedOnWorldMethodWithPayload = wasFaultInjectedOnMethodWherePayloadContains("cloud.filibuster.examples.WorldService/World", request.toString());
+                    assertTrue(wasFaultInjectedOnWorldMethodWithPayload);
+                }
+
+                // Second RPC failed.
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: UNIMPLEMENTED")) {
+                    expected = true;
+                    secondRPCFailed = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: INTERNAL")) {
+                    expected = true;
+                    secondRPCFailed = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: UNAVAILABLE")) {
+                    expected = true;
+                    secondRPCFailed = true;
+                }
+
+                if (t.getMessage().equals("DATA_LOSS: io.grpc.StatusRuntimeException: DATA_LOSS: io.grpc.StatusRuntimeException: DEADLINE_EXCEEDED")) {
+                    expected = true;
+                    secondRPCFailed = true;
+                }
+
+                if (secondRPCFailed) {
+                    boolean wasFaultInjectedOnWorldService = wasFaultInjectedOnService("HelloService");
+                    assertTrue(wasFaultInjectedOnWorldService);
+
+                    boolean wasFaultInjectedOnWorldMethod = wasFaultInjectedOnMethod("cloud.filibuster.examples.HelloService/Hello");
+                    assertTrue(wasFaultInjectedOnWorldMethod);
+                }
 
                 if (!expected) {
                     throw t;
                 }
             } else {
-                throw t;
+                if (!expected) {
+                    throw t;
+                }
             }
         }
 
@@ -106,24 +147,24 @@ public class JUnitFilibusterDataNondeterminismTest extends JUnitBaseTest {
         helloChannel.awaitTermination(1000, TimeUnit.SECONDS);
     }
 
-    @DisplayName("Verify correct number of thrown exceptions.")
+    @DisplayName("Verify correct exceptions thrown.")
     @Test
     @Order(2)
     public void testNumAssertions() {
-        assertEquals(4, testExceptionsThrown.size());
+        assertEquals(8, testExceptionsThrown.size());
     }
 
     @DisplayName("Verify correct number of executed tests.")
     @Test
     @Order(3)
     public void testNumberOfTestsExecuted() {
-        assertEquals(5, numberOfTestsExecuted);
+        assertEquals(9, numberOfTestsExecuted);
     }
 
     @DisplayName("Verify correct number of exceptions thrown.")
     @Test
     @Order(4)
     public void numberOfExceptionsThrown() {
-        assertEquals(4, numberOfExceptionsThrown);
+        assertEquals(8, numberOfExceptionsThrown);
     }
 }
